+++ lib/src/abis/safe4337Module.g.dart
@@ -1,5 +1,8 @@
-// Generated code, do not modify. Run `build_runner build` to re-generate!
-// @dart=2.12
+// @dart=3.0
+// coverage:ignore-file
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
+// ignore_for_file: unused_local_variable, unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
 // ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'package:web3dart/web3dart.dart' as _i1;
 import 'dart:typed_data' as _i2;
@@ -58,20 +61,22 @@ class Safe4337Module extends _i1.GeneratedContract {
   /// like the gas price, nonce and max gas. The `data` and `to` fields will be
   /// set by the contract.
   Future<String> executeUserOp(
-    _i1.EthereumAddress to,
-    BigInt value,
-    _i2.Uint8List data,
-    BigInt operation, {
+    ({
+      _i1.EthereumAddress to,
+      BigInt value,
+      _i2.Uint8List data,
+      BigInt operation
+    }) args, {
     required _i1.Credentials credentials,
     _i1.Transaction? transaction,
   }) async {
     final function = self.abi.functions[3];
     assert(checkSignature(function, '7bb37428'));
     final params = [
-      to,
-      value,
-      data,
-      operation,
+      args.to,
+      args.value,
+      args.data,
+      args.operation,
     ];
     return write(
       credentials,
@@ -85,20 +90,22 @@ class Safe4337Module extends _i1.GeneratedContract {
   /// like the gas price, nonce and max gas. The `data` and `to` fields will be
   /// set by the contract.
   Future<String> executeUserOpWithErrorString(
-    _i1.EthereumAddress to,
-    BigInt value,
-    _i2.Uint8List data,
-    BigInt operation, {
+    ({
+      _i1.EthereumAddress to,
+      BigInt value,
+      _i2.Uint8List data,
+      BigInt operation
+    }) args, {
     required _i1.Credentials credentials,
     _i1.Transaction? transaction,
   }) async {
     final function = self.abi.functions[4];
     assert(checkSignature(function, '541d63c8'));
     final params = [
-      to,
-      value,
-      data,
-      operation,
+      args.to,
+      args.value,
+      args.data,
+      args.operation,
     ];
     return write(
       credentials,
@@ -112,12 +119,12 @@ class Safe4337Module extends _i1.GeneratedContract {
   /// set, the function will be evaluated in the specified block. By default, the
   /// latest on-chain block will be used.
   Future<_i2.Uint8List> getOperationHash(
-    dynamic userOp, {
+    ({dynamic userOp}) args, {
     _i1.BlockNum? atBlock,
   }) async {
     final function = self.abi.functions[5];
     assert(checkSignature(function, 'b25f3776'));
-    final params = [userOp];
+    final params = [args.userOp];
     final response = await read(
       function,
       params,
@@ -130,12 +137,12 @@ class Safe4337Module extends _i1.GeneratedContract {
   /// set, the function will be evaluated in the specified block. By default, the
   /// latest on-chain block will be used.
   Future<_i2.Uint8List> getOperationHash$2(
-    dynamic userOp, {
+    ({dynamic userOp}) args, {
     _i1.BlockNum? atBlock,
   }) async {
     final function = self.abi.functions[6];
     assert(checkSignature(function, 'bbe5dc4f'));
-    final params = [userOp];
+    final params = [args.userOp];
     final response = await read(
       function,
       params,