+++ lib/src/interfaces/account_factories.dart
@@ -4,11 +4,10 @@ part of 'interfaces.dart';
 ///
 /// This class defines the common interface for interacting with an Ethereum smart contract
 /// responsible for creating and managing accounts.
-abstract class SimpleAccountFactoryBase {
+abstract class LightAccountFactoryBase {
   /// Retrieves the Ethereum address associated with a standard account.
   Future<EthereumAddress> getAddress(
-    EthereumAddress owner,
-    BigInt salt, {
+    ({EthereumAddress owner, BigInt salt}) args, {
     BlockNum? atBlock,
   });
 }
@@ -16,8 +15,7 @@ abstract class SimpleAccountFactoryBase {
 abstract class P256AccountFactoryBase {
   /// Retrieves the Ethereum address associated with a standard p256 account.
   Future<EthereumAddress> getP256AccountAddress(
-    BigInt salt,
-    Uint8List creation, {
+    ({BigInt salt, Uint8List creation}) args, {
     BlockNum? atBlock,
   });
 }
@@ -26,9 +24,11 @@ abstract class SafeProxyFactoryBase {
   Future<Uint8List> proxyCreationCode({BlockNum? atBlock});
 
   Future<String> createProxyWithNonce(
-    EthereumAddress singleton,
-    Uint8List initializer,
-    BigInt saltNonce, {
+    ({
+      EthereumAddress singleton,
+      Uint8List initializer,
+      BigInt saltNonce
+    }) args, {
     required Credentials credentials,
     Transaction? transaction,
   });